+++ src/storybook-helpers.ts
@@ -13,7 +13,7 @@ import {
   getMethods,
 } from "./cem-parser.js";
 import { Component, getComponentByTagName } from "@wc-toolkit/cem-utilities";
-import type { ArgTypes } from "./storybook-types";
+import type { ArgTypes } from '@storybook/web-components';
 import type { Categories, Options, StoryHelpers, StoryOptions } from "./types";
 import type { Package } from "custom-elements-manifest";
 
@@ -50,7 +50,7 @@ export function setStorybookHelpersConfig(options: Options) {
  */
 export function getStorybookHelpers<T>(
   tagName: string,
-  options?: StoryOptions
+  options?: StoryOptions,
 ): StoryHelpers<T> {
   userOptions = (globalThis as any)?.__WC_STORYBOOK_HELPERS_CONFIG__ || {};
   const cem = getManifest();
@@ -73,7 +73,7 @@ export function getStorybookHelpers<T>(
         slot,
         argTypes,
         options?.excludeCategories || [],
-        options?.setComponentVariable
+        options?.setComponentVariable,
       ),
   };
 
@@ -85,7 +85,7 @@ function getManifest(): Package {
   const cem: Package = (window as any).__STORYBOOK_CUSTOM_ELEMENTS_MANIFEST__;
   if (!cem) {
     throw new Error(
-      `Custom Elements Manifest not found. Be sure to follow the pre-install steps in this guide:\nhttps://www.npmjs.com/package/wc-storybook-helpers#before-you-install`
+      `Custom Elements Manifest not found. Be sure to follow the pre-install steps in this guide:\nhttps://www.npmjs.com/package/wc-storybook-helpers#before-you-install`,
     );
   }
   return cem;
@@ -96,7 +96,7 @@ function getComponent(cem: Package, tagName: string): Component | undefined {
 
   if (!component) {
     throw new Error(
-      `A component with the tag name "${tagName}" was not found in the Custom Elements Manifest. If it's missing in the CEM, it's often the result of a missing "@tag" or "@tagName" tag in the component's JSDoc.\nAdditional information can be found here:\nhttps://custom-elements-manifest.open-wc.org/analyzer/getting-started/#supported-jsdoc`
+      `A component with the tag name "${tagName}" was not found in the Custom Elements Manifest. If it's missing in the CEM, it's often the result of a missing "@tag" or "@tagName" tag in the component's JSDoc.\nAdditional information can be found here:\nhttps://custom-elements-manifest.open-wc.org/analyzer/getting-started/#supported-jsdoc`,
     );
   }
 
@@ -110,7 +110,7 @@ function getComponent(cem: Package, tagName: string): Component | undefined {
  */
 function getArgTypes(
   component?: Component,
-  excludeCategories?: Array<Categories>
+  excludeCategories?: Array<Categories>,
 ): ArgTypes {
   const cssProps = getCssProperties(component);
   const cssParts = getCssParts(component);
@@ -141,7 +141,7 @@ function getArgTypes(
     attrsAndProps.resets,
     events.resets,
     cssStates.resets,
-    methods.resets
+    methods.resets,
   );
 
   userOptions.categoryOrder?.forEach((category) => {
@@ -158,8 +158,8 @@ function getArgTypes(
  * @param argTypes argTypes object for component
  * @returns an object containing the `args` for the component
  */
-function getArgs<T>(argTypes: ArgTypes): Partial<T> {
-  const args: Partial<T> = {};
+function getArgs<T>(argTypes: ArgTypes): Partial<T> & { [key: string]: any } {
+  const args: Partial<T> & { [key: string]: any } = {};
   for (const [key, value] of Object.entries(argTypes)) {
     if (value?.control) {
       args[key as keyof T] = getDefaultValue(value.defaultValue) || "";
@@ -191,7 +191,7 @@ function getDefaultValue(value?: string | number | boolean | object) {
  */
 function getReactProps(
   component?: Component,
-  excludeCategories?: Array<Categories>
+  excludeCategories?: Array<Categories>,
 ): ArgTypes {
   const cssProps = getCssProperties(component);
   const cssParts = getCssParts(component);
@@ -227,7 +227,7 @@ function getReactProps(
     (category) => {
       if (excludeCategories?.includes(category)) return;
       argTypes = { ...argTypes, ...(args[category] || {}) };
-    }
+    },
   );
 
   return argTypes;