+++ src/types.ts
@@ -1,5 +1,8 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
 import { TemplateResult } from "lit";
-import { ArgTypes } from "./storybook-types";
+import type {
+  ArgTypes
+} from '@storybook/web-components';
 
 export type Categories =
   | "attributes"
@@ -32,11 +35,14 @@ export type StoryOptions = {
 };
 
 export type StoryHelpers<T> = {
-  args: Partial<T>;
+  args: Partial<T> & { [key: string]: any };
   argTypes: ArgTypes;
   reactArgs: Record<string, unknown>;
   reactArgTypes: ArgTypes;
   events: string[];
   styleTemplate: (args?: Record<string, unknown>) => TemplateResult | "";
-  template: (args?: Partial<T>, slot?: TemplateResult) => TemplateResult;
+  template: (
+    args?: Partial<T> & { [key: string]: any },
+    slot?: TemplateResult,
+  ) => TemplateResult;
 };